requirements.txt

langchain == 0.1.1
openai == 1.8.0
langchain_openai == 0.0.2.post1
python-pptx == 0.6.22
qrcode == 7.4.2
pytesseract == 0.3.10
Pillow == 8.1.0

# after installing spacy, run this command to download the english model
# python -m spacy download en_core_web_sm
